FIX : double and int issue edit sign up and sign in views and buttons

+++ lib/features/checkout/presentation/view model/cubit/cart_cubit.dart
@@ -10,7 +10,7 @@ part 'cart_state.dart';
 class CartCubit extends Cubit<CartState> {
   CartCubit() : super(CartInitial());
 
-  double total = 0;
+  int total = 0;
   final CartRepImp cartRepImp = CartRepImp();
   List<CartModel> currentCartList = [];
